+++ cpu/sam0_common/Makefile.features
@@ -12,4 +12,7 @@ FEATURES_PROVIDED += periph_uart_nonblocking
 FEATURES_PROVIDED += periph_uart_reconfigure
 FEATURES_PROVIDED += periph_wdt periph_wdt_cb periph_wdt_warning_period
 
+FEATURES_CONFLICT += periph_rtc:periph_rtt
+FEATURES_CONFLICT_MSG += "The RTC and RTT map to the same hardware peripheral."
+
 -include $(RIOTCPU)/cortexm_common/Makefile.features
